[update] float combat lvl

+++ db/migrate/20180622012554_add_combat_to_players.rb
@@ -1,5 +1,5 @@
 class AddCombatToPlayers < ActiveRecord::Migration
   def change
-    add_column :players, :combat_lvl, :float
+    add_column :players, :combat_lvl, :int
   end
 end
